Fix 4 more verified bugs: robustness, reconnect replay, integrity
- moo: guard JSON.parse of host.chatFactSummaries() so a transient host/DB
  error can't abort the entire chat listing (matches sibling defensive pattern)
- broadcast: remember compaction-draft in ACTIVE_DRAFTS so it's replayed on
  reconnect like draft/reasoning-draft (was silently dropped)
- snapshots: compare file sizes in u64 instead of truncating to usize, so the
  content-addressed size check can't false-match on 32-bit targets
- host: recover-from-poison (lock_recover) in trace_store_put_blob/apply and
  the in-flight write registration, instead of silently dropping the write —
  matches the convention used everywhere else for these mutexes

+++ src/snapshots.rs
@@ -188,7 +188,7 @@ fn write_atomic(path: &Path, bytes: &[u8]) -> Result<(), String> {
         Err(_) if path.exists() => {
             let _ = fs::remove_file(&tmp);
             match fs::metadata(path) {
-                Ok(meta) if meta.len() as usize == bytes.len() => Ok(()),
+                Ok(meta) if meta.len() == bytes.len() as u64 => Ok(()),
                 Ok(_) => Err(format!(
                     "content-addressed path already exists with different size: {}",
                     path.display()
